+++ gt-core/src/entities_frontend/user_info_ts.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 pub struct Model {
     pub id: i32,
     pub user_id: i32,
-    pub height: Option<f32>,
-    pub weight: Option<f32>,
-    pub muscle_mass: Option<f32>,
-    pub body_fat: Option<f32>,
+    pub height: Option<f64>,
+    pub weight: Option<f64>,
+    pub muscle_mass: Option<f64>,
+    pub body_fat: Option<f64>,
     pub created_at: DateTime,
 }
